Piping cleanup, redirections folder created, small libft update
