+++ src/redirections/minishell_redirections_01.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   minishell_redirections_01.c                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jvarila <jvarila@student.hive.fi>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/08 09:59:04 by jvarila           #+#    #+#             */
+/*   Updated: 2025/04/08 09:59:14 by jvarila          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	handle_redirections(t_minishell *data)
+{
+	(void)data;
+}
